==> 00_Import_and_descriptive_analysis.ipynb <==
# 01. Importing libraries

# Import libraries

import pandas as pd
import numpy as np
import os

# 02. Importing data

#### Methode A (pandas)

# import csv within a certain folder 

df = pd.read_csv(r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis\02 Data\021 Original data\orders.csv', index_col = False)

#### Methode B (os)

# using os.path.join() function

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

path

# importing df
df = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'orders.csv'), index_col = False)

# 03. Exploring data with pandas

# display first 5 rows in the df object
df.head()

df.head(10)

df.tail(10)

# check the dimensions of a df; how many rows and columns

df.shape

# basic info about the df

df.info ()

## Interpretation
# pandas df
# 3,421,083 entries (rows) with an index starting from 0 to 3,421,082
# 7 columns
# 5 int64s, 1 float64, and 1 object (=string)
# DF takes up 182.7+ MB of space

# Checking only data types

df.dtypes

# Check columns

df.columns

# generate descriptive statistics

df.describe()

## Other visualization

print(df.describe())

pd.options.display.float_format = '{:.2f}'.format

print(df.describe())

pd.options.display.float_format = None
print(df.describe())

# 04. Lists

# Import only first 1000 rows

df = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'orders.csv'), nrows=1000)

# import a data set as a list, without unnecessary columns

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

# only import the columns from the list

df = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'orders.csv'), usecols = vars_list)

df.head()
