==> 01_Import_ords_prods_and_descriprive.ipynb <==
# Importing libraries

import pandas as pd
import numpy as np
import os

# Orders.csv

# Importing df

df = pd.read_csv(r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis\02 Data\021 Original data\orders.csv', index_col = False)

# Import the “orders.csv” data set into your notebook using the os library

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

path

df = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'orders.csv'), index_col = False)

# exclude variable eval_set
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

# Print the first 5 rows

df.head(5)

# Products.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'products.csv'), index_col = False)

# first 20 rows
df_prods.head(20)

# last 35 rows
df_prods.tail(35)

# Print the names of the columns
df_prods.dtypes

# Print the number of rows and columns
df_prods.shape

# What’s the maximum value of the “aisle_id” column?



df_prods['aisle_id'].max()

# What’s the data type of the “department_id” column?

df_prods['department_id'].dtypes
