==> 02_Data Wrangling_Subsetting.ipynb <==
# Data wrangling

# Importing libraries

import pandas as pd
import numpy as np
import os

# Importing df

df_ords = pd.read_csv(r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis\02 Data\021 Original data\orders.csv', index_col = False)

## 01. Dropping columns

# dropping 'eval_set' column --> only visualization
df_ords.drop(columns = ['eval_set'])

df_ords['days_since_prior_order'].value_counts(dropna = False)

## 02. Renaming columns

## Renaming the column

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head(5)

## 03. Changing a variable's data type

df_ords.describe()

# change the data type
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

## 04. Transposing data

df_dep = pd.read_csv(r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis\02 Data\021 Original data\departments.csv', index_col = False)

df_dep.head(5)

# transpose

df_dep.T

# overwriting the df to a new transposed one

df_dep_t = df_dep.T

df_dep_t

# adding an index to a df

df_dep_t.reset_index()

# using the 1st row for header
# iloc = “locate an integer”

new_header = df_dep_t.iloc[0]

new_header

# creating a new df that contains only rows beyound the first row

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# adding a new header

df_dep_t_new.columns = new_header
df_dep_t_new

## 05. Data dictionaries

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'
df_prods = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'products.csv'), index_col = False)
df_prods.head(3)

## Creating a dictionary

# turn the new dataframe into a dictionary 
# to_dict() function

data_dict = df_dep_t_new.to_dict('index')
data_dict

df_prods.head()

# check what stands behind 19 using a dictionary
print(data_dict.get('19'))

## 06. Subsetting

#Create df_prods subset for snacks department
df_snacks =  df_prods[df_prods['department_id']==19]
df_snacks

# The --- object['element'] --- syntax is called indexing

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks.head()

#### Another way to achive the same result

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

# One more way

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

## 07. Exporting dataframes

df_ords.to_csv(os.path.join(path, '02 Data','022 Prepared Data', 'orders_wrangled.csv'), header=True, index=False)

df_dep_t_new.to_csv(os.path.join(path, '02 Data','022 Prepared Data', 'departments_transposed.csv'), header=True)

df_ords.head()

df_dep_t_new.head()

## Testing

df_dep_new = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'departments_transposed.csv'), index_col = False)

df_dep_new.head()

df_ords_new = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords_new.head()
