==> 03_Data_Wrangling_and_analysis.ipynb <==
## Data Wrangling & Subsetting

import pandas as pd
import numpy as np
import os

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

# Original data Products
df_prods = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'products.csv'), index_col = False)

# Changed data
df_ords = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_wrangled.csv'), index_col = False)
df_deps_t = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'departments_transposed.csv'), index_col = False)

## I used changes csv's from the exercised

df_prods.head()

df_ords.head()

df_deps_t.head()

## 01. Identifier variable in the df_ords dataframe

df_ords.head()

## User_Id to string

df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords['user_id'].dtype

## 02. Looking for a variable in df_ords dataframe with an unintuitive name and changing its name without overwriting the data frame.

## Renaming the column

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)
df_ords.head()

## 03. What the busiest hour is for placing orders?

# calculating the frequency

df_ords['order_hour_of_day'].value_counts(dropna = True)

#### Answer: 10 hours is a busiest hour for placing orders (has the highest count = 288418)

## 04. The meaning behind a value of 4 in the "department_id" column within the df_prods dataframe

## Creating a dictionary
data_dict = df_deps_t.to_dict('records')
data_dict

# check what stands behind 4 using a dictionary
# print(data_dict.get('4'))
# can't get the answer with the code from the lecture, so I tried to find another way

print(data_dict[3])

## One more way

# define the new variable with the value 4
department_id_to_find = 4

# define the variable with the name of the department
# use the function next() to return the first matching value
# chect if the 'department_id' in that dictionary matches the value of department_id_to_find (4)
# return "None" is nothing is found
department_info = next((item for item in data_dict if item['department_id'] == department_id_to_find), None)

# Print output
print(department_info)

# 05. Breakfast sales

## Create a subset containing only the required information > "Breakfast"

df_breakfast = df_prods.loc[df_prods['department_id']==14]
df_breakfast.head()

# 06. All observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood.

## Create a subset containing only the required information

df_dinner = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]
df_dinner

# 07. How many rows does the last dataframe you created have?

num_rows = df_dinner.shape[0]
print("Number of rows in df_dinner:", num_rows)

# 08. Information about user_id = 1

user_id_1 = df_ords.loc[df_ords['user_id']==1]
user_id_1

# using readable values format
pd.options.display.float_format = '{:.2f}'.format
# convert oder_id to string
df_ords['order_id'] = df_ords['order_id'].astype(str)
# statistics about user_id = 1
user_id_1.describe()

#  09. What basic stats can you provide based on the information you have?

#### Answer: 
# The customer with user_id = 1 has made 11 orders, with most orders placed at the beginning of the week on Tuesdays and Wednesdays. 
# The customer tends to place orders around 11 AM, with the earliest order placed at 7 AM and the latest order at 4 PM. 
# Average time between orders is 19 days.
