==> 04_Consistency_Checks_Ords_Prods.ipynb <==
# Data Consistency Checks

import pandas as pd
import numpy as np
import os

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'products.csv'), index_col = False)
df_ords = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_wrangled.csv'), index_col = False)

## 01. Data quality

## Dimensions
# reliability, accuracy, consistency, completeness, uniqueness, and timeliness

df_ords.describe()

# Check percentile "orders_day_of_week"
# Range have to be 0-6 = 7 days a week

var_days = df_ords['orders_day_of_week']

print("50th percentile of orders_day_of_week : ",
       np.percentile(var_days, 50))
print("25th percentile of orders_day_of_week : ",
       np.percentile(var_days, 25))
print("75th percentile of orders_day_of_week : ",
       np.percentile(var_days, 75))

# Check percentile "order_hour_of_day"
# It have to be max 24 hours

var_hours = df_ords['order_hour_of_day']

print("50th percentile of order_hour_of_day : ",
       np.percentile(var_hours, 50))
print("25th percentile of order_hour_of_day : ",
       np.percentile(var_hours, 25))
print("75th percentile of order_hour_of_day : ",
       np.percentile(var_hours, 75))

# Check percentile "days_since_prior_order"
# It can't be negative

var_days_since = df_ords['days_since_prior_order']

print("50th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 50))
print("25th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 25))
print("75th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 75))

# Check for missing values in the 'days_since_prior_order' column
missing_values = df_ords['days_since_prior_order'].isnull().sum()
missing_values

## Issue: Missing values (206209) in the variable 'days_since_prior_order'

## 02. Mixed-Type Data

# mixed-type column is a column that includes both string values and numeric values

## Working with an created dataframe
#
# Creating a dataframe

df_test = pd.DataFrame()

# Creating a mixed type column

df_test['mix'] = ['a', 'b', 1, True] 

df_test.head()

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

## Explanation
# for-loop > for this elements do this + loop describes how
# weird > new variable, which will take a boolean valeu of True or False
## Where
# True > column contains inconsistent data types
# False > contains only data type
## If
# if the conditions is met, do that
# Here: execute a line of code
## This code check all columns in the df for problematic and print their names

#### How to fix mix columns

df_test['mix'] = df_test['mix'].astype('str')

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'].dtype

## 03. Missing values

# Finding missing values

df_prods.isnull().sum()

df_prods.isnull()

# Creating a new df, which contains only missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

## Case with the non-string values:
# Check df
df_prods.shape

df_prods['product_name'].isnull().sum()

# Creating a new df without missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean.isnull().sum()

df_prods_clean.shape

## Command to drop all missing values
# df_prods.dropna(inplace = True)
## Dropping only the NaNs from a particular column
# df_prods.dropna(subset = [‘product_name’], inplace = True)

## 04. Duplicates

## Finding duplicates
# Looking for full duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

df_prods_clean.shape

## Addressing Duplicates
# Deleting (pandas function)
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

## 05. Tidying Up and Exporting Changes

# Exporting to a new csv with the name products_checked.csv

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','022 Prepared Data', 'products_checked.csv'), header=True, index=False)

## 06. Testing

df_prods_test = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'products_checked.csv'), index_col = False)

# Checking for mix

for col in df_prods_test.columns:
  weird_test = (df_prods_test[[col]].applymap(type) != df_prods_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods_test[weird_test]) > 0:
    print (col)

# Checking for missing values

df_prods_test.isnull().sum()

# Checking for duplicates

df_dups_test = df_prods_test[df_prods_test.duplicated()]
df_dups_test

df_prods_test.head()
