==> 05_Addit6ional_Consistency_Checks.ipynb <==
## Data Consistency Checks

import pandas as pd
import numpy as np
import os 

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_prods_check = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'products_checked.csv'), index_col = False)

## 01.Consistence check of df_prods

df_prods_check.describe()

df_prods_check['prices'].median()

--> The maximal price value (99999.00) seems to be too large or incorrectly recorded. Client's verification needed

## 02. Investigate a df and share in a markdown cell whether anything about the data looks off or should be investigated further

pd.options.display.float_format = '{:.2f}'.format
df_ords.describe()

Comments:
1. Order ID > according to the table, there should be 3421083 orders
2. User ID > In total 206209 users
3. Order Number > In total 100 orders made
4. Orders_day_of_week > from 0 (Monday) to 6 (Sunday) --> 7 days a week
5. Hours a day > from 0 to 23 --> 24 houts a day
6. Days_since_prior_order > from 0 to 30

df_ords.head()

# The column Days_since_prior_order contains missing (NaN) values
# Checking the column in separate

# Check percentile "days_since_prior_order"
# It can't be negative

var_days_since = df_ords['days_since_prior_order']

print("50th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 50))
print("25th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 25))
print("75th percentile of days_since_prior_order : ",
       np.percentile(var_days_since, 75))

# Check for missing values in the 'days_since_prior_order' column
missing_values = df_ords['days_since_prior_order'].isnull().sum()
missing_values

## Issue: Missing values (206209) in the variable 'days_since_prior_order'

## 03. Checking for mixed-type data

for col in df_ords.columns:
  mix_test = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[mix_test]) > 0:
    print (col)

--> No output for this code means that there is no mixed-type data

# Checking data types
df_ords.dtypes

## 04. Running a check for missing values

df_ords.isnull().sum()

df_ords['days_since_prior_order'].describe()

df_ords.shape

# How many percent of data is missed

percent_of_missed_data = (206209/3214874)*100
print("{:.2f}%".format(percent_of_missed_data))

--> The amount of missing values is not large enough to significantly bias the results of the analysis (6.41%). From this point on, the missing values can be removed or changed to the average, median or interpolate values.

## 05. Adressing the missing values

df_ords['days_since_prior_order'].median()


# Changing missed values to median 
# Creating a new cleaned df

df_ords_clean = df_ords.copy()
median_days = df_ords_clean['days_since_prior_order'].median()
df_ords_clean['days_since_prior_order'].fillna(median_days, inplace=True)

# Checking the results
df_ords_clean.isnull().sum()

df_ords_clean.head()

--> I chose the method of filling missing values with the median because the percentage of missing values is relatively small (6.41%), which indicates that the impact of these missing values on the analysis results would be minimal. 

--> Using the median, a robust central tendency measure, helps maintain the overall distribution and reduces the potential bias that could arise from using more complex imputation methods for this dataset.

## 06. Running a check for duplicate values

df_dups = df_ords_clean[df_ords_clean.duplicated()]

df_dups.head()

## 07. Exporting the cleaned dataframes

df_ords_clean.to_csv(os.path.join(path, '02 Data','022 Prepared Data', 'orders_checked.csv'), header=True, index=False)
