==> 06_Combining_Merging.ipynb <==
## Combining & Exporting Data

# Import libraries

import pandas as pd
import numpy as np
import os

## 01. Methods for combining data

### Merging the Instacart Data

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_checked.csv'), index_col = False)

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', '021 Original Data', 'order_products__prior.csv'), index_col = False)

# Check the output

df_ords_prior.head()

df_ords.head()

df_ords_prior.shape

df_ords.shape

# Merging the dfs
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)
df_merged_large.head()

df_merged_large['_merge'].value_counts()

--> Conclusion (full match here) is wrong

# Merging the dfs --> TEST
df_merged_large_outer = df_ords.merge(df_ords_prior, on = 'order_id', how = 'outer', indicator = True)
df_merged_large_outer['_merge'].value_counts() 

### Exporting Data in Pickle Format

# Export data to csv

df_merged_large.to_csv(os.path.join(path, '02 Data','022 Prepared Data', 'orders_products_combined.csv'))

# Export data to pkl

df_merged_large.to_pickle(os.path.join(path, '02 Data','022 Prepared Data', 'orders_products_combined.pkl'))

#### Testing:

df_merged_large = pd.read_pickle(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_products_combined.pkl'))

df_merged_large.head()
