==> 07_Combining_Exporting_Data.ipynb <==
## Combining & Exporting Data

# Import libraries

import pandas as pd
import numpy as np
import os 

# Import the orders_products_combined

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'
df_ords_merged = pd.read_pickle(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_products_combined.pkl'))
df_prods_check = pd.read_csv(os.path.join(path, '02 Data', '022 Prepared Data', 'products_checked.csv'), index_col = False)

## 01. Checking the shape

# Previous shape -> df_ords_prior.shape == (32434489, 4) & df_ords.shape == (3421083, 7)
df_ords_merged.shape

## 02. Determine a way to combine with products data set

df_ords_merged.head()

df_prods_check.head()

df_prods_check.shape

pd.options.display.float_format = '{:.2f}'.format
df_ords_merged['product_id'].describe()

df_prods_check['product_id'].describe()

--> Both datasets have a common variable "product_id". This variable could be used to merge the data.

## Merging the data
# 1. Drop the "_merge" column (otherwise: ValueError: Cannot use name of an existing column for indicator column)
df_ords_merged = df_ords_merged.drop(columns = ['_merge'])
df_ords_merged.head()

# 2. Merge the data sets
df_prods_ords_merged = df_ords_merged.merge(df_prods_check, on = 'product_id', indicator = True)

df_prods_ords_merged.head()

## 03. Confirming the results of the merge using the merge flag

df_prods_ords_merged['_merge'].value_counts()

# check the shape and compate to previous --> df_ords_merged.shape (32434489, 11) & df_prods_check.shape (49672, 5)

df_prods_ords_merged.shape

df_merged_large_outer = df_ords_merged.merge(df_prods_check, on = 'product_id', how = 'outer', indicator = True)
df_merged_large_outer['_merge'].value_counts() 

## 04. Exporting in Pickle Format

# Export data to pkl

df_prods_ords_merged.to_pickle(os.path.join(path, '02 Data','022 Prepared Data', 'orders_products_merged.pkl'))

## 05. Testing

df_prods_ords_merged = pd.read_pickle(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_products_merged.pkl'))

df_prods_ords_merged.head()

df_prods_ords_merged.shape
