==> 09_Grouping_Aggregating.ipynb <==
## Grouping Data & Aggregating Variables

# Importing libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os

## 01. Grouping data

# Importing data frames

path = r'C:\Users\efens\cf_tasks\2023-07 Instacard Basket Analysis'
ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', '022 Prepared Data', 'orders_products_merged.pkl'))

# Creating a subset for the first one million entires

df = ords_prods_merged[:1000000]

df.head()

df.shape

#### Grouping Data with pandas

## Using the groupby() function

df.groupby('product_name')

## 02. Performing a Single Aggregation

Calculating the mean of the “order_number” column grouped by the “department_id” column

- Split the data into groups based on department_id
- Apply the agg() function to each group to obtain the mean values for the “order_number” column

df.groupby('department_id').agg({'order_number': ['mean']})

# testing on the whole data frame

ords_prods_merged.groupby('department_id').agg({'order_number': ['mean']})

# calculating the mean value for each department

df.groupby('department_id')['order_number'].mean()

# other way

df.groupby('department_id').order_number.mean()

## 03. Performing Multiple Aggregations

# Executing other statistical valued from the variable "order_number"

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

#### Aggregating Data with transform()

Creating a flag for loyal customers:

1. max orders > 40 --> Loyal customer
2. max orders > 10 and <= 40 --> Regular customer
3. max orders < 10 --> New customer

Steps for creating: 

1. Split the data into groups based on the “user_id” column
2. Apply the transform() function on the “order_number” column to generate the maximum orders for each user
3. Create a new column, “max_order,” into which you’ll place the results of your aggregation

# all steps from above using a single code line

ords_prods_merged['max_order'] = ords_prods_merged.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merged['max_order']

ords_prods_merged.head(15)

ords_prods_merged.head(100)

# Removing output limit from 100 to none

pd.options.display.max_rows = None

## 04. Deriving Columns with loc()

# max orders > 40 --> Loyal customer

ords_prods_merged.loc[ords_prods_merged['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# max orders > 10 and <= 40 --> Regular customer

ords_prods_merged.loc[(ords_prods_merged['max_order'] <= 40) & (ords_prods_merged['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# max orders < 10 --> New customer

ords_prods_merged.loc[ords_prods_merged['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merged['loyalty_flag'].value_counts(dropna = False)

# Checking

ords_prods_merged[['user_id', 'loyalty_flag', 'order_number']].head(60)

ords_prods_merged.head()

# Exclude the eval_set column

ords_prods_merged = ords_prods_merged.loc[:, ords_prods_merged.columns != 'eval_set'] 

ords_prods_merged.head()

# Exporting aggregated df

ords_prods_merged.to_pickle(os.path.join(path, '02 Data','022 Prepared Data', 'ords_prods_merged_agg_df.pkl'))
